Remove isCli method
